Fix accoriding to feedbacks
1. Move RuleAttemptsListener to base class AbstractRelOptPlanner so it can be used in Hep Planner as well
2. Fix style issue with import
