+++ core/src/main/java/org/apache/calcite/plan/hep/HepPlanner.java
@@ -190,6 +190,11 @@ public class HepPlanner extends AbstractRelOptPlanner {
     // Get rid of everything except what's in the final plan.
     collectGarbage();
 
+    if (LOGGER.isDebugEnabled()) {
+      LOGGER.debug("Hep Planner Rule Attempts Info:");
+      LOGGER.debug(dumpRuleAttemptsInfo());
+    }
+
     return buildFinalPlan(root);
   }
